\name{export}
\alias{export}
\alias{mosek}
\alias{mosek_clean}
\alias{mosek_version}
\alias{mosek_read}
\alias{mosek_write}
\title{Unknown exported object}
\description{
  Unknown exported object. Please use 'Rmosek::mosek_attachbuilder' to complete the installation of Rmosek.
}
\seealso{
\code{\link{mosek_attachbuilder}}
}
