\name{export}
\alias{export}
\alias{mosek}
\alias{mosek_clean}
\alias{mosek_version}
\alias{mosek_read}
\alias{mosek_write}
\title{Unknown exported objects to be built}
\description{
  Unknown exported objects to be built:
  \itemize{
    \item \code{mosek}
    \item \code{mosek_clean}
    \item \code{mosek_version}
    \item \code{mosek_read}
    \item \code{mosek_write}
  }
  Please use 'Rmosek::mosek_attachbuilder' to complete the installation of Rmosek.
}
\seealso{
\code{\link{mosek_attachbuilder}}
}
